refactor: replace widget-toolbar to widgets folder
