+++ frontend/src/app/features/dashboard/dashboard.ts
@@ -4,7 +4,7 @@ import { GridsterConfig, GridsterModule, GridType } from 'angular-gridster2';
 import { Subscription } from 'rxjs';
 import { Widget } from '../../interfaces/widget.interface';
 import { WidgetService } from '../../services/widget.service';
-import { WidgetToolbar } from '../widget-toolbar/widget-toolbar';
+import { WidgetToolbar } from '../widgets/widget-toolbar/widget-toolbar';
 import { ZoomService } from '../../services/zoom.service';
 
 @Component({
