+++ frontend/src/app/features/widgets/widget-toolbar/widget-toolbar.html
